Modificacion del adaptador de home. Agregué las canciones a la base de datos y cambié el color de la app, ya debemos empezar con la logica para que reproduzca pero no se como jajaj
