+++ app/src/main/res/values/colors.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="colorPrimary">#6200EE</color>
-    <color name="colorPrimaryDark">#3700B3</color>
+    <color name="colorPrimary">#64B5FC</color>
+    <color name="colorPrimaryDark">#64B5FC</color>
     <color name="colorAccent">#03DAC5</color>
     <color name="color_letters">#000000</color>
 </resources>
